refactor: test config 패키징
